api/requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
MarkupSafe==2.1.2
pycparser==2.21
pydantic==1.10.8
PyJWT==2.7.0
PyMySQL==1.0.3
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.27.0
typing_extensions==4.6.2
uvicorn==0.22.0
Werkzeug==2.3.4
